+++ handlers/kinetic_request_ce_submission_db_insert_v1/test/simple_input.rb
@@ -2,26 +2,26 @@
   'info' => {
     'jdbc_database_id' => 'postgresql',
     'jdbc_url_opts' => '',
-    'host' => 'localhost',
-    'port' => '5432',
-    'database_name' => 'submission_reporting',
+    'host' => '',
+    'port' => '54320',
+    'database_name' => 'my_postgres',
     'database_identifier_size' => '',
-    'user' => 'postgres',
-    'password' => 'postgres',
+    'user' => '',
+    'password' => '',
     'enable_debug_logging' => 'true',
-    'api_server' => 'https://request.domain.com',
-    'api_username' => 'test',
-    'api_password' => 'test',
-    'space_slug' => 'kinetic-data',
+    'api_server' => '',
+    'api_username' => '',
+    'api_password' => '',
+    'space_slug' => '',
     'enable_debug_logging'=>'yes'
   },
   'parameters' => {
     'error_handling' => 'Error Message',
-    'space_slug' => 'kinetic-data',
+    'space_slug' => '',
     'kapp_slug' => 'services',
-    'form_slug' => 'general-support-request',
+    'form_slug' => 'cleaning',
     'datastore' => '',
-    'submission_id' => "a6c3114d-302e-11e7-a02c-912ba6cf418f",
+    'submission_id' => '',
     'submission_deletion_timestamp' => ''
   }
 }