+++ changelog.md
@@ -61,3 +61,6 @@ Kinetic Platform [handlers] (2021-10-30)
   * [kinetic_request_ce_attachment_copy_remote_server_v1]
     * initial commit.
 
+Kinetic Platform [handlers (2022-04-01)
+  * [kinetic_request_ce_submission_db_insert_v1]
+    * PER-268 fixed unique key constraint bug.